Simplify blade Timoshenko element creation
